refactor: rename appUtils.ts to app-util.ts for kebab-case consistency
Rename the file to use kebab-case to match the project's file naming
convention. Update all import statements and comments that reference
the old filename.
