+++ source/hooks/useAppHandlers.tsx
@@ -1,5 +1,7 @@
-import {handleMessageSubmission} from '@/app/utils/appUtils';
-import {createClearMessagesHandler} from '@/app/utils/appUtils';
+import {
+	createClearMessagesHandler,
+	handleMessageSubmission,
+} from '@/app/utils/app-util';
 import {
 	ErrorMessage,
 	SuccessMessage,
